Update bnb for new repo layout

+++ cryodrgn/dataset.py
@@ -49,7 +49,7 @@ class LazyMRCData(data.Dataset):
         assert ny == nx, "Images must be square"
         assert ny % 2 == 0, "Image size must be even"
         log('Loaded {} {}x{} images'.format(N, ny, nx))
-        self.particles = particles
+        self.particles = particles.astype(np.float32)
         self.N = N
         self.D = ny + 1 # after symmetrizing HT
         self.invert_data = invert_data
